feat(types): bootstrap reddb-io-types keystone crate, re-home type vocabulary (#1061)
Bootstrap the neutral keystone crate reddb-io-types (ADR 0052) at the
bottom of the workspace crate graph — zero dependencies on other
workspace crates. Re-home the logical type vocabulary (Value, DataType,
SqlTypeName, TypeModifier, TypeCategory, ValueError, Row) byte-faithfully
from reddb-server's storage::schema::types, together with its sibling
value_codec module (Value::to_bytes/from_bytes delegate to it, and the
embedded #[cfg(test)] tests round-trip through it). Both moved files are
byte-identical to their former server locations.

Server keeps re-export shims at storage::schema::{types,value_codec} so
the ~180 existing call-sites stay untouched. value_codec carries an
external zstd (TOAST) dep only; no workspace-crate dependency is
introduced. The keystone crate's 56 embedded tests pass in-crate.
